services/api/requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
psycopg2-binary==2.9.9
pydantic==2.9.2
